+++ packages/platforms/browser/rollup.config.npm.mjs
@@ -2,5 +2,5 @@ import fs from 'fs'
 import createRollupConfig from '../../../.rollup/index.mjs'
 
 export default createRollupConfig({
-  external: ['@bugsnag/cuid'],
+  external: ['@bugsnag/cuid', '@bugsnag/delivery-fetch-performance'],
 })